Basic Ubuntu Xenial Kea Build
This is a first stab at building the latest version of kea from git.
I should likely move to only building stable tags. The version available
in xenial right now is 1.0.0 which is far out of date from looking at the
docs which have many mentions of breaks in 1.1 and 1.2.