Add Archlinux Kea Image
This builds an archlinux kea image that can also be used. I originally
thought that it would be much smaller than the one I build using ubuntu
however it's actually 10MB larger... Likely easier to maintain in the longrun
though but i'll leave the old version around incase the community doesn't keep
kea up to date in the arch repos.