Fixus Ghost url/env
Make sure we are running ghost in production mode only as I'm using it
for a blog and don't care to do any js development locally. Added in an
example of how to set the url so it won't use the ip:port mapping of
the container making the website unusable.